(PC-35429) feat(ui): split AsyncErrorBoundaryWithoutNavigation and AsyncErrorBoundary
