+++ src/features/errors/pages/AsyncErrorBoundary.native.test.tsx
@@ -2,6 +2,7 @@ import React, { useEffect } from 'react'
 import { ErrorBoundary } from 'react-error-boundary'
 
 import { ApiError } from 'api/ApiError'
+import { AsyncErrorBoundary } from 'features/errors/pages/AsyncErrorBoundary'
 import { useMaintenance } from 'features/maintenance/helpers/useMaintenance/useMaintenance'
 import { MaintenanceErrorPage } from 'features/maintenance/pages/MaintenanceErrorPage'
 import * as useGoBack from 'features/navigation/useGoBack'
@@ -10,8 +11,6 @@ import { DEFAULT_REMOTE_CONFIG } from 'libs/firebase/remoteConfig/remoteConfig.c
 import { AsyncError, MonitoringError, ScreenError, LogTypeEnum } from 'libs/monitoring/errors'
 import { eventMonitoring } from 'libs/monitoring/services'
 import { fireEvent, render, screen } from 'tests/utils'
-
-import { AsyncErrorBoundary } from './AsyncErrorBoundary'
 
 jest.mock('libs/monitoring/services')
